+++ packages/api-codegen-preset/src/api-project.ts
@@ -1,7 +1,6 @@
-import "@shopify/hydrogen-codegen/patch";
 import fs from "fs";
 
-import { pluckConfig } from "@shopify/hydrogen-codegen";
+import { pluckConfig } from "@shopify/graphql-codegen";
 
 import type { ShopifyApiProjectOptions } from "./types";
 import { shopifyApiTypes } from "./api-types";