+++ packages/graphql-client/src/graphql-client/graphql-client.ts
@@ -19,6 +19,10 @@ import {
   HEADER_SEPARATOR,
   DEFER_OPERATION_REGEX,
   BOUNDARY_HEADER_REGEX,
+  SDK_VARIANT_HEADER,
+  SDK_VERSION_HEADER,
+  DEFAULT_SDK_VARIANT,
+  DEFAULT_CLIENT_VERSION,
 } from "./constants";
 import {
   formatErrorMessage,
@@ -122,6 +126,11 @@ function generateFetch(
       return headers;
     }, {});
 
+    if (!flatHeaders[SDK_VARIANT_HEADER] && !flatHeaders[SDK_VERSION_HEADER]) {
+      flatHeaders[SDK_VARIANT_HEADER] = DEFAULT_SDK_VARIANT;
+      flatHeaders[SDK_VERSION_HEADER] = DEFAULT_CLIENT_VERSION;
+    }
+
     const fetchParams: Parameters<CustomFetchApi> = [
       overrideUrl ?? url,
       {
@@ -188,14 +197,15 @@ function generateRequest(
 async function* getStreamBodyIterator(
   response: Response,
 ): AsyncIterableIterator<string> {
-  // Support node-fetch format
+  const decoder = new TextDecoder();
+
+  // Response body is an async iterator
   if ((response.body as any)![Symbol.asyncIterator]) {
     for await (const chunk of response.body! as any) {
-      yield (chunk as Buffer).toString();
+      yield decoder.decode(chunk);
     }
   } else {
     const reader = response.body!.getReader();
-    const decoder = new TextDecoder();
 
     let readResult: ReadableStreamReadResult<DataChunk>;
     try {