+++ packages/graphql-client/src/graphql-client/tests/graphql-client/fixtures.ts
@@ -4,7 +4,17 @@ import { Readable } from "stream";
 import { ReadableStream } from "web-streams-polyfill/es2018";
 
 import { createGraphQLClient } from "../../graphql-client";
-import { LogContentTypes, ClientOptions } from "../../types";
+import {
+  LogContentTypes,
+  ClientOptions,
+  Headers as TypesHeaders,
+} from "../../types";
+import {
+  SDK_VARIANT_HEADER,
+  SDK_VERSION_HEADER,
+  DEFAULT_CLIENT_VERSION,
+  DEFAULT_SDK_VARIANT,
+} from "../../constants";
 
 global.TextEncoder = TextEncoder;
 global.TextDecoder = TextDecoder as any;
@@ -29,12 +39,20 @@ export const variables = {
   country: "US",
 };
 
+export const defaultHeaders = {
+  ...clientConfig.headers,
+  [SDK_VARIANT_HEADER]: DEFAULT_SDK_VARIANT,
+  [SDK_VERSION_HEADER]: DEFAULT_CLIENT_VERSION,
+};
+
 export function getValidClient({
   retries,
   logger,
+  headers,
 }: {
   retries?: number;
   logger?: (logContent: LogContentTypes) => void;
+  headers?: TypesHeaders;
 } = {}) {
   const updatedConfig: ClientOptions = { ...clientConfig };
 
@@ -44,6 +62,13 @@ export function getValidClient({
 
   if (logger !== undefined) {
     updatedConfig.logger = logger;
+  }
+
+  if (headers !== undefined) {
+    updatedConfig.headers = {
+      ...updatedConfig.headers,
+      ...headers,
+    };
   }
 
   return createGraphQLClient(updatedConfig);
@@ -57,9 +82,17 @@ const streamResponseConfig = {
   }),
 };
 
+function getStringEncoder() {
+  const textEncoder = new TextEncoder();
+
+  return (str: any) => {
+    return textEncoder.encode(str);
+  };
+}
+
 function createReadableStream(
   responseArray: string[],
-  stringEncoder?: (str: any) => Uint8Array,
+  stringEncoder: (str: any) => Uint8Array = getStringEncoder(),
 ) {
   return new ReadableStream({
     start(controller) {
@@ -67,7 +100,7 @@ function createReadableStream(
       queueData();
       function queueData() {
         const chunk = responseArray[index];
-        const string = stringEncoder ? stringEncoder(chunk) : chunk;
+        const string = stringEncoder(chunk);
 
         // Add the string to the stream
         controller.enqueue(string);
@@ -86,10 +119,7 @@ function createReadableStream(
 }
 
 export function createReaderStreamResponse(responseArray: string[]) {
-  const encoder = new TextEncoder();
-  const stream = createReadableStream(responseArray, (str) => {
-    return encoder.encode(str);
-  });
+  const stream = createReadableStream(responseArray);
 
   return {
     ...streamResponseConfig,
@@ -101,6 +131,15 @@ export function createReaderStreamResponse(responseArray: string[]) {
 
 export function createIterableResponse(responseArray: string[]) {
   const stream = createReadableStream(responseArray);
+
+  return new Response(Readable.from(stream) as any, streamResponseConfig);
+}
+
+export function createIterableBufferResponse(responseArray: string[]) {
+  const encoder = new TextEncoder();
+  const stream = createReadableStream(responseArray, (str) => {
+    return Buffer.from(encoder.encode(str));
+  });
 
   return new Response(Readable.from(stream) as any, streamResponseConfig);
 }