+++ packages/shopify-api/future/flags.ts
@@ -1,3 +1,6 @@
+import {type ShopifyLogger} from '../lib/logger';
+import {type ConfigInterface} from '../lib/base-types';
+
 /**
  * Future flags are used to enable features that are not yet available by default.
  */
@@ -21,3 +24,22 @@ export type FeatureEnabled<
     ? true
     : false
   : false;
+
+export function logDisabledFutureFlags(
+  config: ConfigInterface,
+  logger: ShopifyLogger,
+) {
+  if (!config._logDisabledFutureFlags) {
+    return;
+  }
+
+  const logFlag = (flag: string, message: string) =>
+    logger.info(`Future flag ${flag} is disabled.\n\n  ${message}\n`);
+
+  if (!config.future?.v10_lineItemBilling) {
+    logFlag(
+      'v10_lineItemBilling',
+      'Enable this flag to use the new billing API, that supports multiple line items per plan.',
+    );
+  }
+}