+++ packages/admin-api-client/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @shopify/admin-api-client
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies [e9652b7]
+  - @shopify/graphql-client@0.10.3
+
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies [2f862e3]
+- Updated dependencies [b2f29ae]
+  - @shopify/graphql-client@0.10.2
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [56d2fcd]
+  - @shopify/graphql-client@0.10.1
+
 ## 0.2.5
 
 ### Patch Changes