+++ packages/api-codegen-preset/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.0.5
+
+### Patch Changes
+
+- 8e06306: Replaced @shopify/hydrogen-codegen with its more generic successor, @shopify/graphql-codegen. All of the changes are internal, so this package is unchanged.
+
+## 0.0.4
+
+### Patch Changes
+
+- 37f1b7b: Updated dependency on @shopify/hydrogen-codegen
+
 ## 0.0.3
 
 ### Patch Changes