+++ packages/graphql-client/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @shopify/graphql-client
 
+## 0.10.3
+
+### Patch Changes
+
+- e9652b7: Remove `Partial` around the `ClientStreamResponse.data` type
+
+## 0.10.2
+
+### Patch Changes
+
+- 2f862e3: Update `requestStream()` to always process response iterator chunks as `Uint8Array` values
+- b2f29ae: Add default SDK headers to all API requests when none are provided
+
+## 0.10.1
+
+### Patch Changes
+
+- 56d2fcd: Update `requestStream()` to support `Uint8Array` streamed encoded format in `node` environments
+
 ## 0.10.0
 
 ### Minor Changes