+++ packages/storefront-api-client/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @shopify/storefront-api-client
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies [e9652b7]
+  - @shopify/graphql-client@0.10.3
+
+## 0.3.2
+
+### Patch Changes
+
+- Updated dependencies [2f862e3]
+- Updated dependencies [b2f29ae]
+  - @shopify/graphql-client@0.10.2
+
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [56d2fcd]
+  - @shopify/graphql-client@0.10.1
+
 ## 0.3.0
 
 ### Minor Changes